determine is this porsea project or not
